feat (tests): Refactor testing suite
